added proxy for UI and BE connection, refactored code
