+++ FE/src/components/feature/Genre/types/genre.interface.ts
@@ -1,0 +1,4 @@
+export interface Genre {
+  id: string;
+  name: string;
+}
