--- FE/src/constants/routes.jsx
@@ -1,7 +1,0 @@
-export const ROUTES = {
-  HOME: '/',
-  ERROR: '*',
-  ADMIN: {
-    GENRES: '/admin-panel/genres'
-  }
-}
