+++ FE/src/constants/routes.ts
@@ -1,0 +1,7 @@
+export const ROUTES = {
+  HOME: "/",
+  ERROR: "*",
+  ADMIN: {
+    GENRES: "/admin-panel/genres",
+  },
+};
