+++ FE/src/pages/Home/HomePage.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import { ROUTES } from './../../constants/routes';
+import { ROUTES } from '../../constants/routes';
 import "./HomePage.scss";
 
 const HomePage = () => {
